Correct compiler warnings

+++ tapeccws.c
@@ -3208,9 +3208,9 @@ BYTE            rustat;                 /* Addl CSW stat on Rewind Unload */
 
             /* Copy device filename to guest storage */
             if ( dev->filename != NULL )
-                str_host_to_guest( (const BYTE*)dev->filename, iobuf, num );
+                str_host_to_guest( (const BYTE*)dev->filename, iobuf, (u_int)num );
             else
-                str_host_to_guest( "", iobuf, num );
+                str_host_to_guest( "", iobuf, (u_int)num );
 
             /* Return normal status */
             build_senseX (TAPE_BSENSE_STATUSONLY, dev, unitstat, code);
